Changed the attr onModelUpdate of Cart
The name of the onModelUpdate attr of cartwas changed to afterModelUpdate,
not the function, to afterModelUpdate, since it created some conflict
when calling the function on the model cycle. The value binds to the attr were
also modified in facade.js

Also, the new distribution files are included in the commit.